Fix output printing problems
This was a twofold problem.

First, there was a problem with the indentation in `run_jneuroml`,
which prevented printing of output (and in some cases even proper
return from the function).

Second, global variables in Python cannot be used as default parameter
values. So the use of the `DEFAULTS` dictionary had to be changed.